Add spec for ContentfulMigrations::Migration

+++ lib/contentful_migrations/migration.rb
@@ -2,17 +2,18 @@ module ContentfulMigrations
   class Migration
     attr_reader :name, :version, :contentful_client, :contentful_space
 
-    def initialize(name = self.class.name, version = nil)
+    def initialize(name = self.class.name, version = nil, client = nil, space = nil)
       @name       = name
       @version    = version
-      @contentful_client = nil
-      @contentful_space = nil
+      @contentful_client = client
+      @contentful_space = space
     end
 
     def migrate(direction, client, space)
       @contentful_client = client
       @contentful_space = space
       send(direction)
+      self
     end
 
     def with_space
@@ -32,8 +33,10 @@ module ContentfulMigrations
 
     def erase_migration(migration_content_type)
       entry = migration_content_type.entries.all.find { |m| m.version.to_i == version.to_i }
+      return unless entry
       entry.unpublish
       entry.destroy
+      entry
     end
   end
 end
